test: move hidden-fields outbound test to memora-internal
The test was using the base memora destination where the fields are not
declared, so the framework stripped them trivially. Moved to memora-internal
where segment_traits_by_field and segment_identifiers_by_field are declared,
making the assertion meaningful.
